move markdown regex to regex file, rename index to mapTSfiles
