+++ pkgs/model/src/demos/generateCode/generator.js
@@ -1,4 +1,4 @@
-import { getMapData } from "./index.js"
+import { getMapData } from "./mapTSfiles.js"
 import { replaceMarkupCode } from "./markup.js"
 
 /**
