+++ pkgs/model/src/demos/generateCode/regex.js
@@ -3,5 +3,7 @@ export const regex = {
     blockCommentMatch:
         /\/\*\*( ){0,}\n?( ){0,}\*( ){0,}\@re_place( ){1,}\w+\.(js|ts)(( ){1,}(\w+))?( ){0,}\n?( ){0,}\*\//g,
     commentMatch:
-        /\/\/( ){0,}@re_place( ){1,}[\w]+[\.](js|ts)(( ){1,}([\w]+))?/g
+        /\/\/( ){0,}@re_place( ){1,}[\w]+[\.](js|ts)(( ){1,}([\w]+))?/g,
+    markdownTagMatch:
+        /([*]{3}GENERATED[*]{3}( ){1,})?( ){0,}[\w]+\.(ts|js)(( ){1,}([\w]+))?/g
 }
